+++ src/bump_scope_guard.rs
@@ -1,10 +1,8 @@
 use core::{fmt::Debug, marker::PhantomData, num::NonZeroUsize, ptr::NonNull};
 
-use allocator_api2::alloc::Allocator;
-
 use crate::{
-    chunk_header::ChunkHeader, polyfill::nonnull, Bump, BumpScope, Chunk, GuaranteedAllocatedStats, MinimumAlignment,
-    RawChunk, Stats, SupportedMinimumAlignment,
+    chunk_header::ChunkHeader, polyfill::nonnull, BaseAllocator, Bump, BumpScope, Chunk, GuaranteedAllocatedStats,
+    MinimumAlignment, RawChunk, Stats, SupportedMinimumAlignment,
 };
 
 /// This is returned from [`checkpoint`](Bump::checkpoint) and used for [`reset_to`](Bump::reset_to).
@@ -31,7 +29,7 @@ impl Checkpoint {
 pub struct BumpScopeGuard<'a, A, const MIN_ALIGN: usize = 1, const UP: bool = true>
 where
     MinimumAlignment<MIN_ALIGN>: SupportedMinimumAlignment,
-    A: Allocator + Clone,
+    A: BaseAllocator,
 {
     pub(crate) chunk: RawChunk<UP, A>,
     address: usize,
@@ -41,7 +39,7 @@ where
 impl<'a, A, const MIN_ALIGN: usize, const UP: bool> Debug for BumpScopeGuard<'a, A, MIN_ALIGN, UP>
 where
     MinimumAlignment<MIN_ALIGN>: SupportedMinimumAlignment,
-    A: Allocator + Clone,
+    A: BaseAllocator,
 {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         self.stats().debug_format("BumpScopeGuard", f)
@@ -51,7 +49,7 @@ where
 impl<'a, A, const MIN_ALIGN: usize, const UP: bool> Drop for BumpScopeGuard<'a, A, MIN_ALIGN, UP>
 where
     MinimumAlignment<MIN_ALIGN>: SupportedMinimumAlignment,
-    A: Allocator + Clone,
+    A: BaseAllocator,
 {
     #[inline(always)]
     fn drop(&mut self) {
@@ -62,7 +60,7 @@ where
 impl<'a, A, const MIN_ALIGN: usize, const UP: bool> BumpScopeGuard<'a, A, MIN_ALIGN, UP>
 where
     MinimumAlignment<MIN_ALIGN>: SupportedMinimumAlignment,
-    A: Allocator + Clone,
+    A: BaseAllocator,
 {
     #[inline(always)]
     #[allow(clippy::needless_pass_by_ref_mut)]
@@ -126,7 +124,7 @@ where
 pub struct BumpScopeGuardRoot<'b, A, const MIN_ALIGN: usize = 1, const UP: bool = true>
 where
     MinimumAlignment<MIN_ALIGN>: SupportedMinimumAlignment,
-    A: Allocator + Clone,
+    A: BaseAllocator,
 {
     pub(crate) chunk: RawChunk<UP, A>,
     marker: PhantomData<&'b mut ()>,
@@ -135,7 +133,7 @@ where
 impl<'b, A, const MIN_ALIGN: usize, const UP: bool> Debug for BumpScopeGuardRoot<'b, A, MIN_ALIGN, UP>
 where
     MinimumAlignment<MIN_ALIGN>: SupportedMinimumAlignment,
-    A: Allocator + Clone,
+    A: BaseAllocator,
 {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         self.guaranteed_allocated_stats().debug_format("BumpScopeGuardRoot", f)
@@ -145,7 +143,7 @@ where
 impl<'a, A, const MIN_ALIGN: usize, const UP: bool> Drop for BumpScopeGuardRoot<'a, A, MIN_ALIGN, UP>
 where
     MinimumAlignment<MIN_ALIGN>: SupportedMinimumAlignment,
-    A: Allocator + Clone,
+    A: BaseAllocator,
 {
     #[inline(always)]
     fn drop(&mut self) {
@@ -156,7 +154,7 @@ where
 impl<'a, A, const MIN_ALIGN: usize, const UP: bool> BumpScopeGuardRoot<'a, A, MIN_ALIGN, UP>
 where
     MinimumAlignment<MIN_ALIGN>: SupportedMinimumAlignment,
-    A: Allocator + Clone,
+    A: BaseAllocator,
 {
     #[inline(always)]
     #[allow(clippy::needless_pass_by_ref_mut)]